+++ run.py
@@ -1,2 +1,2 @@
-from web import app
+from scoring_engine.web import app
 app.run()
