+++ scoring_engine/models.bak/user.py
@@ -1,0 +1,38 @@
+import bcrypt
+
+from sqlalchemy import Boolean, Column, Integer, String, ForeignKey
+from sqlalchemy.orm import relationship
+
+from models.base import Base
+
+
+class User(Base):
+    __tablename__ = 'users'
+    id = Column(Integer, primary_key=True)
+    username = Column(String(20), nullable=False)
+    password = Column(String(50), nullable=False)
+    team_id = Column(Integer, ForeignKey('teams.id'))
+    team = relationship("Team", back_populates="users")
+    authenticated = Column(Boolean, default=False)
+
+    def check_hash(self, hashed, password):
+        if bcrypt.hashpw(password, hashed) == hashed:
+            return True
+        else:
+            return False
+
+    def is_active(self):
+        """True, as all users are active."""
+        return True
+
+    def get_id(self):
+        """Return the id to satisfy Flask-Login's requirements."""
+        return self.id
+
+    def is_authenticated(self):
+        """Return True if the user is authenticated."""
+        return self.authenticated
+
+    def is_anonymous(self):
+        """False, as anonymous users aren't supported."""
+        return False
