+++ scoring_engine/web/static/images/favicon.ico
