+++ scoring_engine/web/templates/admin.html
