+++ scoring_engine/web/templates/scoreboard.html
