+++ scoring_engine/web/views/welcome.py
